Small portability changes to HTML/CSS
In deployment script, use of the package Beautiful Soup for HTML editing
results in mangling of whitespace between tags.  In particular, having
one character in <b> tags for boldface results in extra whitespace being
inserted, which causes an unsightly space between the letter and the
rest of the word (e.g. the "B" in "Build Tool").  Highlighting
individual characters in HTML text is not part of the design of HTML and
CSS, with the exception of a few CSS pseudo-classes such as
::first-letter, which would work for our purposes, but only is allowed
in block formatted elements, of which input labels are not.

The change will not be a long-term fix, since eventually radio boxes
will be replaced with an icon-formatted tool selector, but for now the
slightly less appealing look will be fine.